moved tiff includes to include/sls
