+++ slsDetectorCalibration/tiffio/CMakeLists.txt
@@ -1,4 +1,4 @@
 find_package(TIFF REQUIRED)
 add_library(tiffio STATIC tiffIO.cpp)
-target_include_directories(tiffio PUBLIC .)
+target_include_directories(tiffio PUBLIC include/)
 target_link_libraries(tiffio PUBLIC TIFF::TIFF)
